correction 1

+++ app/Http/Controllers/PlateController.php
@@ -4,54 +4,51 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\PlateModel;
+use App\Http\Requests\PlateRequest;
+
 class PlateController extends Controller
 {
-    public function AddPlate(Request $req){
-        $validatedData = $req->validate([
-            'name' => 'required|max:255',
-            'description' => 'required',
-        ],
-        [
-            'name.required' => 'Поле Название обязательно для заполнения',
-            'description.required'  => 'Поле Описание обязательно для заполнения',
-        ]
-        );
+    public function AddPlate(PlateRequest $req)
+    {
         $plate = new PlateModel();
         $plate->name = $req->input('name');
         $plate->description = $req->input('description');
         $plate->save();
         return redirect()->route('welcome')->with('success', 'Новая пластинка добавлена');
     }
-    public function deletePlate($id, Request $req){
-        $plate = PlateModel::find($id);
-        $plate->delete();
-        return redirect()->route('welcome')->with('success', 'Пластинка удалена');
+
+    public function deletePlate(Request $req)
+    {   $id=$req->input('id');
+        if($plate = PlateModel::find($id))
+        {
+            $plate->delete();
+            return redirect()->route('welcome')->with('success', 'Пластинка удалена');
+        } else return redirect()->route('welcome')->with('success', 'Пластинки не существует');    
     }
-    public function allDate() {
-       // $plate = DB::table('plate_models')->paginate(15);
+
+    public function getPage() 
+    {
        $plate = new PlateModel();
         return view('welcome', ['data'=>$plate->paginate(5)]);
-        
     }
-    public function editDate($id, Request $req) {
-        $plate= new PlateModel();
-        return view('editplate', ['data'=>$plate->find($id)]);
-        
+
+    public function getPlate($id) 
+    {   
+        if($plate = PlateModel::find($id))
+        {
+            return view('editplate', ['data'=>$plate->find($id)]);
+        } else return redirect()->route('welcome')->with('success', 'Пластинки не существует');
     }
-    public function editPlate($id, Request $req){
-        $validatedData = $req->validate([
-            'name' => 'required',
-            'description' => 'required',
-        ],
-        [
-            'name.required' => 'Поле Название обязательно для заполнения',
-            'description.required'  => 'Поле Описание обязательно для заполнения',
-        ]
-        );
-        $plate = PlateModel::find($id);
-        $plate->name=$req->input('name');
-        $plate->description = $req->input('description');
-        $plate->save();
-        return redirect()->route('welcome')->with('success', 'Редактирование прошло успешно');
+
+    public function editPlate(PlateRequest $req)
+    {   
+        $id=$req->input('id');
+        if($plate = PlateModel::find($id))
+        {
+            $plate->name=$req->input('name');
+            $plate->description = $req->input('description');
+            $plate->save();
+            return redirect()->route('welcome')->with('success', 'Редактирование прошло успешно');
+        } else return redirect()->route('welcome')->with('success', 'Пластинки не существует');     
     }
 }
